Flatten contracts/finp2p/v2 and rename adapter v2 services to orchestration
The v2 subfolder is unnecessary: the new contract file names
(FINP2POrchestrator, FinP2PEscrow, FinP2PPlanVerifier, PlanTypes,
ProofSignerRegistry) don't collide with FINP2POperator /
FINP2POperatorWithRegistry, which are heading for sunset. Adapter-side,
src/services/finp2p-contract-v2 becomes src/services/orchestration, and
the plan-v2 test/config names follow (tests/orchestration.test.ts,
jest.orchestration.config.js, npm run test:orchestration,
deploy-orchestrator script).
