+++ finp2p-contracts/contracts/finp2p/FINP2POrchestrator.sol
@@ -4,11 +4,11 @@ pragma solidity ^0.8.20;
 
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IERC20Metadata} from "@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol";
-import "../../utils/StringUtils.sol";
-import {Burnable} from "../../utils/erc20/Burnable.sol";
-import {Mintable} from "../../utils/erc20/Mintable.sol";
-import {FinP2PReceiptVerifier} from "../../utils/finp2p/FinP2PReceiptVerifier.sol";
-import {FinP2PSignatureVerifier} from "../../utils/finp2p/FinP2PSignatureVerifier.sol";
+import "../utils/StringUtils.sol";
+import {Burnable} from "../utils/erc20/Burnable.sol";
+import {Mintable} from "../utils/erc20/Mintable.sol";
+import {FinP2PReceiptVerifier} from "../utils/finp2p/FinP2PReceiptVerifier.sol";
+import {FinP2PSignatureVerifier} from "../utils/finp2p/FinP2PSignatureVerifier.sol";
 import {FinP2PEscrow} from "./FinP2PEscrow.sol";
 import {FinP2PPlanVerifier} from "./FinP2PPlanVerifier.sol";
 import {ProofSignerRegistry} from "./ProofSignerRegistry.sol";
