+++ finp2p-contracts/package.json
@@ -35,7 +35,7 @@
     "erc20-approve": "ts-node scripts/erc20-approve.ts",
     "eip712-domain-params": "ts-node scripts/eip712-domain-params.ts",
     "finId-to-address": "ts-node scripts/finId-to-address.ts",
-    "deploy-contract-v2": "ts-node scripts/deploy-v2.ts",
+    "deploy-orchestrator": "ts-node scripts/deploy-orchestrator.ts",
     "register-proof-signer": "ts-node scripts/register-proof-signer.ts"
   },
   "keywords": [],
