+++ src/services/orchestration/proof-sync.ts
