+++ tests/orchestration.test.ts
@@ -30,8 +30,8 @@ import {
   PlanEscrowService,
   PlanTokenService,
   ProofSyncService
-} from "../src/services/finp2p-contract-v2";
-import { RawExecutionPlan } from "../src/services/finp2p-contract-v2/plan-translator";
+} from "../src/services/orchestration";
+import { RawExecutionPlan } from "../src/services/orchestration/plan-translator";
 
 const ORG = "bank-us";
 const REMOTE_ORG = "bank-uk";
